changes from original repo (#7)
* FIX: Rebuild fails when tracked `version` switches from branch to tag

We are replacing the `beta` branch with a tag of the same name. Rebuilding a container will fail when a user tracks the `beta` version and the base image was built at a time where `beta` was still a branch. This resolves that by pruning tags and branches and forcing an update of tags during the rebuild.

* FEATURE: Store the `version` parameter as custom value in git config

This is particularly useful when the version is pointing to a git tag like `beta`, otherwise we won't be able to reliably detect the tracked version at runtime.

* DEV: Wait up to 10 minutes for safe PG stop (#709)

Over the years we had LOTS of cases of PostgreSQL instances that weren't stopped properly. This gives PG more time to handle it's shutdown cleanly on the worst case, but will still shutdown fast in normal cases.

---------

Co-authored-by: Gerhard Schlager <gerhard.schlager@discourse.org>
Co-authored-by: Rafael dos Santos Silva <xfalcox@gmail.com>